requirements.txt
cassandra-driver==3.29.3
click==8.3.1
geomet==1.1.0
numpy==2.4.0
pandas==2.3.3
psycopg==3.3.2
psycopg-binary==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
